app/bread-meat-delivery-backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
httpx
python-dotenv
pYJWT
gunicorn